+++ benchmark/benchmark_geofileops.py
@@ -2,14 +2,12 @@ import benchmarker
 
 
 def main():
-    benchmarker.run_benchmarks(["benchmarks_geofileops"])
+    # benchmarker.run_benchmarks(["benchmarks_geofileops"])
 
-    return
+    # return
 
     # Only run specific benchmark function(s)
-    benchmarker.run_benchmarks(
-        ["benchmarks_geofileops"], ["buffer_gridsize_spatialite"]
-    )
+    benchmarker.run_benchmarks(["benchmarks_geofileops"], ["buffer_gpd"])
 
 
 if __name__ == "__main__":