+++ geofileops/fileops.py
@@ -930,6 +930,7 @@ def read_file(
     columns: Optional[Iterable[str]] = None,
     bbox=None,
     rows=None,
+    where: Optional[str] = None,
     sql_stmt: Optional[str] = None,
     sql_dialect: Optional[Literal["SQLITE", "OGRSQL"]] = None,
     ignore_geometry: bool = False,
@@ -970,15 +971,20 @@ def read_file(
             In addition to standard columns, it is also possible
             to specify "fid", a unique index available in all input files. Note that the
             "fid" will be aliased eg. to "fid_1". Defaults to None.
-        bbox ([type], optional): Read only geometries intersecting this bbox.
+        bbox (Tuple, optional): return only geometries intersecting this bbox.
             Defaults to None, then all rows are read.
-        rows ([type], optional): Read only the rows specified.
-            Defaults to None, then all rows are read.
+        rows (slice, optional): return only the rows specified. For many file formats
+            (e.g. Geopackage) this is slow, so using e.g. a where filter instead is
+            recommended. Defaults to None, then all rows are returned.
+        where (str, optional): only return the rows that comply to the filter specified.
+            Filter should be in |restricted_where| sql format. Defaults to None.
         sql_stmt (str): sql statement to use. Only supported with "pyogrio" engine.
-        sql_dialect (str, optional): Sql dialect used. If None, for data sources with
-            explicit SQL support the statement is processed by the default SQL engine
-            (e.g. PostGIS, Geopackage, Spatialite,...). For data sources
-            without SQL support, the "OGRSQL" dialect is the default. Defaults to None.
+        sql_dialect (str, optional): Sql dialect used. Options are None, "SQLITE" or
+            "OGRSQL". If None, for data sources with explicit SQL support the statement
+            is processed by the default SQL engine (e.g. for Geopackage and Spatialite
+            this is "SQLITE"). For data sources without native SQL support (e.g. .shp),
+            the "OGRSQL" dialect is the default. If the "SQLITE" dialect is specified,
+            |spatialite_reference_link| functions can also be used. Defaults to None.
         ignore_geometry (bool, optional): True not to read/return the geometry.
             Defaults to False.
         fid_as_index (bool, optional): If True, will use the FIDs of the features that
@@ -990,13 +996,23 @@ def read_file(
 
     Returns:
         gpd.GeoDataFrame: the data read.
-    """
+
+    .. |restricted_where| raw:: html
+
+        <a href="https://ogdi.sourceforge.net/prop/6.2.CapabilitiesMetadata.html#:~:text=qe_format%3Drestricted_where" target="_blank">restricted where</a>
+
+    .. |spatialite_reference_link| raw:: html
+
+        <a href="https://www.gaia-gis.it/gaia-sins/spatialite-sql-latest.html" target="_blank">spatialite reference</a>
+
+    """  # noqa: E501
     result_gdf = _read_file_base(
         path=path,
         layer=layer,
         columns=columns,
         bbox=bbox,
         rows=rows,
+        where=where,
         sql_stmt=sql_stmt,
         sql_dialect=sql_dialect,
         ignore_geometry=ignore_geometry,
@@ -1046,6 +1062,7 @@ def _read_file_base(
     columns: Optional[Iterable[str]] = None,
     bbox=None,
     rows=None,
+    where: Optional[str] = None,
     sql_stmt: Optional[str] = None,
     sql_dialect: Optional[Literal["SQLITE", "OGRSQL"]] = None,
     ignore_geometry: bool = False,
@@ -1069,6 +1086,7 @@ def _read_file_base(
             columns=columns,
             bbox=bbox,
             rows=rows,
+            where=where,
             sql_stmt=sql_stmt,
             sql_dialect=sql_dialect,
             ignore_geometry=ignore_geometry,
@@ -1081,6 +1099,7 @@ def _read_file_base(
             columns=columns,
             bbox=bbox,
             rows=rows,
+            where=where,
             sql_stmt=sql_stmt,
             sql_dialect=sql_dialect,
             ignore_geometry=ignore_geometry,
@@ -1104,6 +1123,7 @@ def _read_file_base_fiona(
     columns: Optional[Iterable[str]] = None,
     bbox=None,
     rows=None,
+    where: Optional[str] = None,
     sql_stmt: Optional[str] = None,
     sql_dialect: Optional[Literal["SQLITE", "OGRSQL"]] = None,
     ignore_geometry: bool = False,
@@ -1164,6 +1184,7 @@ def _read_file_base_fiona(
         bbox=bbox,
         rows=rows,
         include_fields=columns_list,
+        where=where,
         sql=sql_stmt,
         sql_dialect=sql_dialect,
         ignore_geometry=ignore_geometry,
@@ -1203,6 +1224,7 @@ def _read_file_base_pyogrio(
     columns: Optional[Iterable[str]] = None,
     bbox=None,
     rows=None,
+    where: Optional[str] = None,
     sql_stmt: Optional[str] = None,
     sql_dialect: Optional[Literal["SQLITE", "OGRSQL"]] = None,
     ignore_geometry: bool = False,
@@ -1216,13 +1238,15 @@ def _read_file_base_pyogrio(
     if path.exists() is False:
         raise ValueError(f"file doesn't exist: {path}")
 
-    # Convert slice object to pyogrio parameters
+    # Convert rows slice object to pyogrio parameters
     if rows is not None:
         skip_features = rows.start
         max_features = rows.stop - rows.start
     else:
         skip_features = 0
         max_features = None
+    # Arrow doesn't support filtering rows like this
+    # use_arrow = True if rows is None else False
 
     # If no sql_stmt specified
     columns_prepared = None
@@ -1257,16 +1281,27 @@ def _read_file_base_pyogrio(
         bbox=bbox,
         skip_features=skip_features,
         max_features=max_features,
+        where=where,
         sql=sql_stmt,
         sql_dialect=sql_dialect,
         read_geometry=not ignore_geometry,
         fid_as_index=fid_as_index,
+        # use_arrow=use_arrow,
     )
 
     # Reorder columns + change casing so they are the same as columns parameter
     if columns_prepared is not None and len(columns_prepared) > 0:
         result_gdf = result_gdf[list(columns_prepared) + ["geometry"]]
         result_gdf = result_gdf.rename(columns=columns_prepared)
+
+    # Cast columns that are of object type, but contain datetime.date or datetime.date
+    # to proper datetime64 columns.
+    if len(result_gdf) > 0:
+        for column in result_gdf.select_dtypes(include=["object"]):
+            if isinstance(
+                result_gdf[column].iloc[0], (datetime.date, datetime.datetime)
+            ):
+                result_gdf[column] = pd.to_datetime(result_gdf[column])
 
     assert isinstance(result_gdf, (gpd.GeoDataFrame, pd.DataFrame))
     return result_gdf
@@ -1490,8 +1525,16 @@ def _to_file_fiona(
     # Shapefile doesn't support datetime columns, so first cast them to string
     if GeofileType(path) == GeofileType.ESRIShapefile:
         gdf = gdf.copy()
+        # Columns that have a proper datetime64 type
         for column in gdf.select_dtypes(include=["datetime64"]):
             gdf[column] = gdf[column].astype(str)
+
+        # Columns that are of object type, but contain datetime.date or datetime.date
+        # type data instead of strings.
+        if len(gdf) > 0:
+            for column in gdf.select_dtypes(include=["object"]):
+                if isinstance(gdf[column][0], (datetime.date, datetime.datetime)):
+                    gdf[column] = gdf[column].astype(str)
 
     # Handle some specific cases where the file schema needs to be manipulated.
     schema = None